+++ index.js
@@ -1,41 +1,40 @@
 // Components
-export { AllowedBlocks } from './components/AllowedBlocks';
-export { Background } from './components/Background';
-export { BlockIcons } from './components/BlockIcons';
-export { ColorSettings } from './components/ColorSettings';
-export { ContentSearch } from './components/content-search';
-export { ContentPicker } from './components/ContentPicker';
-export { CustomBlockAppender } from './components/CustomBlockAppender';
-export { DragHandle } from './components/drag-handle';
-export { FetchAllTermSelectControl } from './components/FetchAllTermSelectControl';
-export { HelpWrapper } from './components/HelpWrapper';
-export { IconPicker } from './components/IconPicker';
-export { Image } from './components/Image';
-export { LinkToolbar } from './components/LinkToolbar';
-export { LoadingSpinner } from './components/LoadingSpinner';
-// @todo make index
+// export { AllowedBlocks } from './components/AllowedBlocks';
+// export { Background } from './components/Background';
+// export { BlockIcons } from './components/BlockIcons';
+// export { ColorSettings } from './components/ColorSettings';
+// export { ContentSearch } from './components/content-search';
+// export { ContentPicker } from './components/ContentPicker';
+// export { CustomBlockAppender } from './components/CustomBlockAppender';
+// export { DragHandle } from './components/drag-handle';
+// export { FetchAllTermSelectControl } from './components/FetchAllTermSelectControl';
+// export { HelpWrapper } from './components/HelpWrapper';
+// export { IconPicker } from './components/IconPicker';
+// export { Image } from './components/Image/index.mjs';
+// export { LinkToolbar } from './components/LinkToolbar';
+export { LoadingSpinner } from './components/LoadingSpinner/index.mjs';
 // export { MediaCredit } from './components/MediaCredit/media-credit.js';
-export { Optional } from './components/Optional';
-// @todo make index
+// export { Optional } from './components/Optional';
 // export { ParagraphCaptionStyle } from './components/ParagraphCaptionStyle/paragraph-caption-style.js';
-// @todo make index
 // export { ParagraphEndOfArticleStyle } from './components/ParagraphEndOfArticleStyle/paragraph-end-of-article-style.js';
-export { PlainTextWithLimit } from './components/PlainTextWithLimit';
-export { PostChooser } from './components/PostChooser';
-export { PostPicker } from './components/PostPicker';
-export { Repeater } from './components/Repeater';
-export { RichTextWithLimit } from './components/RichTextWithLimit';
-export { ShareTools } from './components/ShareTools';
-export { StyledComponentContext } from './components/styled-components-context';
-export { TermSelector } from './components/TermSelector';
+// export { PlainTextWithLimit } from './components/PlainTextWithLimit';
+// export { PostChooser } from './components/PostChooser';
+// export { PostPicker } from './components/PostPicker';
+// export { Repeater } from './components/Repeater';
+// export { RichTextWithLimit } from './components/RichTextWithLimit';
+// export { ShareTools } from './components/ShareTools';
+// export { StyledComponentContext } from './components/styled-components-context';
+// export { TermSelector } from './components/TermSelector';
 
 // Hooks
-export { useFilteredList } from './hooks/use-filtered-list';
-export { useIcons } from './hooks/use-icons';
-export { useAllTerms } from './hooks/useAllTerms';
+// A React Hook is a special function that lets you "hook into" React state and lifecycle features from within functional components. Hooks enable stateful logic and side effects within functional components, offering a way to reuse logic across components. Crucially, hooks can only be called inside React functional components or custom hooks, and they must follow the "rules of hooks," such as being called at the top level of a component and not within loops or conditional statements.
+// export { useFilteredList } from './hooks/use-filtered-list';
+// export { useIcons } from './hooks/use-icons';
+// export { useAllTerms } from './hooks/useAllTerms';
 export { useMedia } from './hooks/useMedia/index.mjs';
-export { useRenderAppenderWithBlockLimit } from './hooks/useRenderAppenderWithBlockLimit';
-export { useRequestData } from './hooks/useRequestData';
+// export { useRenderAppenderWithBlockLimit } from './hooks/useRenderAppenderWithBlockLimit';
+// export { useRequestData } from './hooks/useRequestData';
 
 // Utils
-export { parseMedia } from './utils/parseMedia/index.mjs';
+// A utility function is a standard JavaScript function that performs a specific task and is not tied to any particular framework or library. Utility functions are often used for tasks like data formatting, calculations, or other operations that don't require access to React's state or lifecycle. They can be called from anywhere in your code, including within React components or hooks.
+export { slugify } from './utils/slugify/index.mjs';
