+++ utils/slugify/README.md
@@ -1,0 +1,10 @@
+# `slugify`
+
+The `slugify` utility is a simple function to convert any string into an one that only uses lowercase letters and hyphens to separate words.
+
+## Usage
+
+```js
+import { slugify } from '@bostonuniversity/block-imports';
+let mySlug = slugify('Jimmy Carson loves peanuts!');
+```
